consortia, consortium, consortium-page all in same file, mapstatetoprops addressed
