+++ packages/coinstac-ui/app/render/components/consortium/consortium-results.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { connect } from 'react-redux';
 import PropTypes from 'prop-types';
 import { Alert } from 'react-bootstrap';
-import { setExpandedResults } from '../../state/ducks/consortia-page';
+import { setExpandedResults } from '../../state/ducks/consortia';
 
 import ConsortiumResult from './consortium-result';
 
@@ -115,7 +115,7 @@ ConsortiumResults.compareRemoteResults = (
   return 0;
 };
 
-const mapStateToProps = ({ consortiaPage: { expandedResults } }) => {
+const mapStateToProps = ({ consortiaState: { expandedResults } }) => {
   return { expandedResults };
 };
