+++ packages/coinstac-ui/app/render/state/root-reducer.js
@@ -2,8 +2,7 @@ import { combineReducers } from 'redux';
 import auth from './ducks/auth';
 import computation from './ducks/computation';
 import computations from './ducks/computations';
-import consortia from './ducks/consortia';
-import consortiaPage from './ducks/consortia-page';
+import consortiaState from './ducks/consortia';
 import loading from './ducks/loading';
 import project from './ducks/project';
 import projects from './ducks/projects';
@@ -14,8 +13,7 @@ const rootReducer = combineReducers({
   auth,
   computation,
   computations,
-  consortia,
-  consortiaPage,
+  consortiaState,
   form,
   loading,
   project,
